+++ src/em_nextion.cpp
@@ -10,54 +10,54 @@ EmNextion::EmNextion(EmComSerial& serial,
                      uint32_t timeoutMs, 
                      EmLogLevel logLevel)
  : EmLog("Nex", logLevel),
-   m_Serial(serial),
-   m_TimeoutMs(timeoutMs),
-   m_IsInit(false)
+   m_serial(serial),
+   m_timeoutMs(timeoutMs),
+   m_isInit(false)
 {
 }
 
-bool EmNextion::Init() const
+bool EmNextion::begin() const
 {
     // Have command feedback on both success/fail  
-    _sendCmdParam("bkcmd=3");
-    _sendCmdEnd();
-    m_IsInit = _ack(ACK_CMD_SUCCEED);
-    return m_IsInit;
+    sendCmdParam_("bkcmd=3");
+    sendCmdEnd_();
+    m_isInit = ack_(ACK_CMD_SUCCEED);
+    return m_isInit;
 }
 
-bool EmNextion::_sendCmd(const char* firstCmd, ...) const
+bool EmNextion::sendCmd_(const char* firstCmd, ...) const
 {
     // Before sending let's see if display is active/connected
-    if (!m_IsInit && !Init()) {
+    if (!m_isInit && !begin()) {
         return false;
     }
-    m_Serial.flush();
-    _sendCmdParam(firstCmd);
+    m_serial.flush();
+    sendCmdParam_(firstCmd);
     va_list args;
     va_start(args, firstCmd);     
     const char* cmdParam = va_arg(args, const char*);
     while (cmdParam) {
-        _sendCmdParam(cmdParam);
+        sendCmdParam_(cmdParam);
         cmdParam = va_arg(args, const char*);
     }
     va_end(args);
 
-    return _sendCmdEnd();
+    return sendCmdEnd_();
 }
 
-bool EmNextion::_sendCmdParam(const char* cmdParam) const
+bool EmNextion::sendCmdParam_(const char* cmdParam) const
 {
-    return _bResult(m_Serial.write(cmdParam) > 0);
+    return bResult_(m_serial.write(cmdParam) > 0);
 }
 
-bool EmNextion::_sendCmdEnd() const
+bool EmNextion::sendCmdEnd_() const
 {
-    m_Serial.write(0xFF);
-    m_Serial.write(0xFF);
-    return _bResult(m_Serial.write(0xFF) == 1);
+    m_serial.write(0xFF);
+    m_serial.write(0xFF);
+    return bResult_(m_serial.write(0xFF) == 1);
 }
 
-EmGetValueResult EmNextion::_recv(uint8_t ackCode, 
+EmGetValueResult EmNextion::recv_(uint8_t ackCode, 
                                   char* buf, 
                                   uint8_t len, 
                                   bool isText) const
@@ -67,10 +67,10 @@ EmGetValueResult EmNextion::_recv(uint8_t ackCode,
     bool got_buffer=(len==0);
     uint8_t term_count=0;
     uint8_t buf_pos=0;
-    EmTimeout rxTimeout(m_TimeoutMs);
-    while (!rxTimeout.IsElapsed(false)) {
-        while (m_Serial.available()) {
-            uint8_t c = static_cast<uint8_t>(m_Serial.read());
+    EmTimeout rxTimeout(m_timeoutMs);
+    while (!rxTimeout.isElapsed(false)) {
+        while (m_serial.available()) {
+            uint8_t c = static_cast<uint8_t>(m_serial.read());
             // Still waiting for ack code?
             if (!got_ackCode) {
                 got_ackCode = (c==ackCode);
@@ -101,24 +101,24 @@ EmGetValueResult EmNextion::_recv(uint8_t ackCode,
                         buf[buf_pos] = 0;
                         continue;
                     }
-                    return _result(false, value_changed);
+                    return result_(false, value_changed);
                 }
                 term_count++;
                 if (term_count>=3) {
                     // Got everything
-                    return _result(true, value_changed);
+                    return result_(true, value_changed);
                 }
             }
         }
     }
-    LogDebug<50>("RX: %s [Timeout elapsed!]", buf);
-    return _result(false, value_changed);
+    logDebug<50>("RX: %s [Timeout elapsed!]", buf);
+    return result_(false, value_changed);
 }
 
-EmGetValueResult EmNextion::_result(bool result, bool valueChanged) const
+EmGetValueResult EmNextion::result_(bool result, bool valueChanged) const
 { 
     if (!result) { 
-        m_IsInit = false;
+        m_isInit = false;
         return EmGetValueResult::failed;
     }
     return valueChanged ? 
@@ -126,63 +126,63 @@ EmGetValueResult EmNextion::_result(bool result, bool valueChanged) const
            EmGetValueResult::succeedEqualValue;
 }
 
-bool EmNextion::_bResult(bool result) const
+bool EmNextion::bResult_(bool result) const
 { 
     if (!result) { 
-        m_IsInit = false;
+        m_isInit = false;
     }
     return result;
 }
 
-bool EmNextion::_ack(uint8_t ackCode) const 
+bool EmNextion::ack_(uint8_t ackCode) const 
 {
-    LogDebug(F("Waiting ACK"));
-    return EmGetValueResult::failed != _recv(ackCode, NULL, 0);
+    logDebug(F("Waiting ACK"));
+    return EmGetValueResult::failed != recv_(ackCode, NULL, 0);
 }
 
-bool EmNextion::IsCurPage(uint8_t pageId) const {
+bool EmNextion::isCurPage(uint8_t pageId) const {
     uint8_t id;
-    return GetCurPage(id) && id == pageId;
+    return getCurPage(id) && id == pageId;
 }
 
-bool EmNextion::GetCurPage(uint8_t& pageId) const 
+bool EmNextion::getCurPage(uint8_t& pageId) const 
 {
-    if (!_sendCmd("sendme", NULL)) {
+    if (!sendCmd_("sendme", NULL)) {
         return false;
     }
-    if (EmGetValueResult::failed != _recv(ACK_CURRENT_PAGE_ID, 
+    if (EmGetValueResult::failed != recv_(ACK_CURRENT_PAGE_ID, 
                                          (char*)&pageId, 1)) {
         return true;
     }
     return false;
 }
 
-bool EmNextion::SetCurPage(uint8_t pageId) const 
+bool EmNextion::setCurPage(uint8_t pageId) const 
 {
     char buf[3];
-    if (!_sendCmd("page ", to_str(buf, 3, pageId), NULL)) {
+    if (!sendCmd_("page ", to_str(buf, 3, pageId), NULL)) {
         return false;
     }
-    return _ack(ACK_CMD_SUCCEED);
+    return ack_(ACK_CMD_SUCCEED);
 }
 
-bool EmNextion::SetCurPage(const char* pageName) const 
+bool EmNextion::setCurPage(const char* pageName) const 
 {
-    if (!_sendCmd("page ", pageName, NULL)) {
+    if (!sendCmd_("page ", pageName, NULL)) {
         return false;
     }
-    return _ack(ACK_CMD_SUCCEED);
+    return ack_(ACK_CMD_SUCCEED);
 }
 
-EmGetValueResult EmNextion::GetNumElementValue(const char* pageName, 
+EmGetValueResult EmNextion::getNumElementValue(const char* pageName, 
                                                const char* elementName, 
                                                int32_t& val) const 
 {
     EmGetValueResult res = EmGetValueResult::failed;
-    if (_sendGetCmd(pageName, elementName, "val")) {
-        res = _getNumber(val);
+    if (sendGetCmd_(pageName, elementName, "val")) {
+        res = getNumber_(val);
     }
-    LogDebug<50>("get: %s -> %d [%s]", 
+    logDebug<50>("get: %s -> %d [%s]", 
                  elementName,
                  val,
                  (EmGetValueResult::failed != res ? 
@@ -191,114 +191,114 @@ EmGetValueResult EmNextion::GetNumElementValue(const char* pageName,
     return res;
 }
 
-bool EmNextion::SetNumElementValue(const char* pageName, 
+bool EmNextion::setNumElementValue(const char* pageName, 
                                    const char* elementName, 
                                    int32_t val) const {
     bool res = false;
-    if (_sendSetCmd(pageName, elementName, "val", val)) {
-        res = _ack(ACK_CMD_SUCCEED);
+    if (sendSetCmd_(pageName, elementName, "val", val)) {
+        res = ack_(ACK_CMD_SUCCEED);
     }
-    LogDebug<50>("set: %s -> %d [%s]", 
+    logDebug<50>("set: %s -> %d [%s]", 
                  elementName,
                  val,
                  (res ? " [SUCCESS]" : " [FAIL]"));
     return res;
 }
 
-bool EmNextion::SetTextElementValue(const char* pageName, 
+bool EmNextion::setTextElementValue(const char* pageName, 
                                     const char* elementName, 
                                     const char* txt) const {
     bool res = false;
-    if (_sendSetCmd(pageName, elementName, "txt", txt)) {
-        res = _ack(ACK_CMD_SUCCEED);
+    if (sendSetCmd_(pageName, elementName, "txt", txt)) {
+        res = ack_(ACK_CMD_SUCCEED);
     }
-    LogDebug<50>("set: %s -> %s [%s]", 
+    logDebug<50>("set: %s -> %s [%s]", 
                  elementName,
                  txt,
                  (res ? " [SUCCESS]" : " [FAIL]"));
     return res;
 }
 
-bool EmNextion::SetVisible(const char* elementName, 
+bool EmNextion::setVisible(const char* elementName, 
                            bool visible) const {
     bool res = false;
-    if (_sendCmd("vis ", elementName, visible ? ",1" : ",0", NULL)) {
-        res = _ack(ACK_CMD_SUCCEED);
+    if (sendCmd_("vis ", elementName, visible ? ",1" : ",0", NULL)) {
+        res = ack_(ACK_CMD_SUCCEED);
     }
-    LogDebug<50>("visible: %s -> %s [%s]", 
+    logDebug<50>("visible: %s -> %s [%s]", 
                  elementName,
                  visible,
                  (res ? " [SUCCESS]" : " [FAIL]"));
     return res;
 }
 
-bool EmNextion::SetVisible(uint8_t pageId, 
+bool EmNextion::setVisible(uint8_t pageId, 
                            const char* elementName, 
                            bool visible) const {
-    return IsCurPage(pageId) && SetVisible(elementName, visible);
+    return isCurPage(pageId) && setVisible(elementName, visible);
 }
 
-bool EmNextion::SetPicture(const char* pageName, 
+bool EmNextion::setPicture(const char* pageName, 
                            const char* elementName, 
                            uint8_t picId) const {
     bool res = false;
-    if (_sendSetCmd(pageName, elementName, "pic", picId)) {
-        res = _ack(ACK_CMD_SUCCEED);
+    if (sendSetCmd_(pageName, elementName, "pic", picId)) {
+        res = ack_(ACK_CMD_SUCCEED);
     }
-    LogDebug<50>("pic: %s -> %s [%s]", 
+    logDebug<50>("pic: %s -> %s [%s]", 
                  elementName,
                  picId,
                  (res ? " [SUCCESS]" : " [FAIL]"));
     return res;
 }
 
-bool EmNextion::GetPicture(const char* pageName, 
+bool EmNextion::getPicture(const char* pageName, 
                            const char* elementName, 
                            uint8_t& picId) const {
     bool res = false;
-    if (_sendGetCmd(pageName, elementName, "val")) {
+    if (sendGetCmd_(pageName, elementName, "val")) {
         int32_t val;
-        res = _getNumber(val) != EmGetValueResult::failed;
+        res = getNumber_(val) != EmGetValueResult::failed;
         if (res) {
             picId = static_cast<uint8_t>(val);
         }
     }
-    LogDebug<50>("pic: %s -> %d [%s]", 
+    logDebug<50>("pic: %s -> %d [%s]", 
                  elementName,
                  picId,
                  res ? " [SUCCESS]" : " [FAIL]");
     return res;
 }
 
-bool EmNextion::Click(const char* elementName, 
+bool EmNextion::click(const char* elementName, 
                       bool pressed) const {
     bool res = false;
-    if (_sendCmd("click ", elementName, pressed ? ",1" : ",0", NULL)) {
-        res = _ack(ACK_CMD_SUCCEED);
+    if (sendCmd_("click ", elementName, pressed ? ",1" : ",0", NULL)) {
+        res = ack_(ACK_CMD_SUCCEED);
     }
-    LogDebug<50>("click: %s -> %s [%s]", 
+    logDebug<50>("click: %s -> %s [%s]", 
                  elementName,
                  pressed,
                  (res ? " [SUCCESS]" : " [FAIL]"));
     return res;
 }
 
-bool EmNextion::Click(uint8_t pageId, 
+bool EmNextion::click(uint8_t pageId, 
                       const char* elementName, 
                       bool pressed) const {
-    return IsCurPage(pageId) && Click(elementName, pressed);
+    return isCurPage(pageId) && click(elementName, pressed);
 }
 
 
-bool EmNextion::_setColor(const char* pageName, 
+bool EmNextion::setColor_(const char* pageName, 
                           const char* elementName, 
                           const char* colorCode, 
                           uint16_t color565) const{
     bool res = false;
-    if (_sendSetCmd(pageName, elementName, colorCode, color565)) {
-        res = _ack(ACK_CMD_SUCCEED);
+    if (sendSetCmd_(pageName, elementName, colorCode, color565)) {
+        res = ack_(ACK_CMD_SUCCEED);
     }
-    LogDebug<50>("%s: %s -> %s [%s]", 
+    logDebug<50>("%s: %s -> %s [%s]", 
                  colorCode,
                  elementName,
                  color565,
@@ -306,19 +306,19 @@ bool EmNextion::_setColor(const char* pageName,
     return res;
 }
 
-bool EmNextion::_getColor(const char* pageName, 
+bool EmNextion::getColor_(const char* pageName, 
                           const char* elementName, 
                           const char* colorCode, 
                           uint16_t& color565) const {
     bool res = false;
-    if (_sendGetCmd(pageName, elementName, colorCode)) {
+    if (sendGetCmd_(pageName, elementName, colorCode)) {
         int32_t val;
-        res = _getNumber(val) != EmGetValueResult::failed;
+        res = getNumber_(val) != EmGetValueResult::failed;
         if (res) {
             color565 = static_cast<uint16_t>(val);
         }
     }
-    LogDebug<50>("%s: %s -> %s [%s]", 
+    logDebug<50>("%s: %s -> %s [%s]", 
                  colorCode,
                  elementName,
                  color565,
@@ -326,46 +326,46 @@ bool EmNextion::_getColor(const char* pageName,
     return res;
 }
 
-bool EmNextion::_sendGetCmd(const char* pageName, 
+bool EmNextion::sendGetCmd_(const char* pageName, 
                   const char* elementName, 
                   const char* property) const
 {
-    return _sendCmd("get ", 
+    return sendCmd_("get ", 
                     pageName, ".", 
                     elementName, ".",
                     property, NULL);
 }
 
-bool EmNextion::_sendSetCmd(const char* pageName, 
+bool EmNextion::sendSetCmd_(const char* pageName, 
                             const char* elementName, 
                             const char* property, 
                             int32_t value) const
 {
     char buf[11];
-    return _sendCmd(pageName, ".", 
+    return sendCmd_(pageName, ".", 
                     elementName, ".",
                     property, "=",
                     to_str(buf, 11, value), NULL);
 }
 
-bool EmNextion::_sendSetCmd(const char* pageName, 
+bool EmNextion::sendSetCmd_(const char* pageName, 
                             const char* elementName, 
                             const char* property, 
                             const char* value) const
 {
-    return _sendCmd(pageName, ".", 
+    return sendCmd_(pageName, ".", 
                     elementName, ".",
                     property, "=",
                     "\"", value, "\"", NULL);
 }
 
 
-EmGetValueResult EmNextion::_getNumber(int32_t& val) const 
+EmGetValueResult EmNextion::getNumber_(int32_t& val) const 
 {
     // Create a copy in case communication fails 
     // (i.e. some bytes might be modified by _recv method!)
     int32_t buf = val;
-    EmGetValueResult res = _recv(ACK_NUMBER, (char*)&buf, sizeof(buf));
+    EmGetValueResult res = recv_(ACK_NUMBER, (char*)&buf, sizeof(buf));
 
     if (EmGetValueResult::failed != res) {
         // TODO: Nextion is little endian, should we check about big endian CPU?
@@ -374,17 +374,17 @@ EmGetValueResult EmNextion::_getNumber(int32_t& val) const
     return res;
 }
 
-EmGetValueResult EmNextion::_getString(char* txt, 
+EmGetValueResult EmNextion::getString_(char* txt, 
                                        uint8_t bufLen, 
                                        const char* elementName) const  
 {
-    EmGetValueResult res = _recv(ACK_STRING, txt, bufLen, true);
+    EmGetValueResult res = recv_(ACK_STRING, txt, bufLen, true);
     if (EmGetValueResult::failed == res) {
         txt[0]=0;
     } else { 
         txt[bufLen-1]=0;
     } 
-    LogDebug<50>("get: %s -> %s [%s]", 
+    logDebug<50>("get: %s -> %s [%s]", 
                  elementName,
                  txt,
                  (EmGetValueResult::failed != res ?